[BUGFIX] Avoid excluding fields in copyToLanguage
Given that a TCA column is using 'l10n_mode' => 'exclude' and creating
a "free localization" of that record using DataHandler's copyToLanguage
command, the according field stays empty in the created localization.

The expected behavior is that the accordings field is still duplicated
once on creating the duplicate for a particular language. Title prefix
"[Translate to <language>:]" is expected to be missing here since it
is the result of a copy command instead of a localize command.

Resolves: #86203
Releases: master
Change-Id: Ib9914a8d02cbbabdbdc56720a983b6314a34709b
Reviewed-on: https://review.typo3.org/58235
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Jürgen Heym <juergen.heym@hof-university.de>
Reviewed-by: Susanne Moog <susanne.moog@typo3.org>
Tested-by: Susanne Moog <susanne.moog@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>