[FEATURE] Add support for the "canonical" tag
TYPO3 will now out-of-the box create the canonical tag if you have
installed the core extension SEO. The canonical tag is based on the
settings in the page properties. First it will check if you defined
a canonical link. If not, it will check for the Show content from page
and if both are not set, it will generate an absolute url to the
current page.

Resolves: #20051
Releases: master
Change-Id: I1ae01e419a1f822b8659bf71ba99523051a2e77c
Reviewed-on: https://review.typo3.org/58271
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Richard Haeser <richard@maxserv.com>
Tested-by: Richard Haeser <richard@maxserv.com>
Reviewed-by: Frank Naegler <frank.naegler@typo3.org>
Tested-by: Frank Naegler <frank.naegler@typo3.org>