[TASK] Deprecate config.tx_extbase.objects and plugin.tx_%name%.objects
If we want to introduce a system wide container in v10,
the non deterministic behaviour of object (re)configurations
during runtime in extbase will be a problem.
Compiled containers like symfony DI do not allow runtime
reconfigurations.

Also the concept of per plugin overwrites is broken anyway,
and should therefore be dropped:
The "objects" configured through plugin.tx_%name%.objects
are are never reset, that means:
Plugin A might overwrite an object and plugin B will implictly use
that overriden class, despite not having configured that at all.
(depending on the fact whether plugin A is rendered before or not)

config.tx_extbase.object class overwrites can be replaced by XCLASSes.
If config.tx_extbase.object was/is used to conditionally overwrite
classes (through sys_template or conditions), these
occasions should be changed to XCLASSes as well. In that case the
conditional overwrite needs to be moved into the xclass itself.

Change-Id: I2ef839bcbb7f941c3afa0010767560d20e15f98b
Releases: master
Resolves: #86270
Reviewed-on: https://review.typo3.org/58288
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>