[TASK] Remove superfluous indention level in site configuration yaml
The site configuration is in folder "sites" in a file called config and
does not need a "site" key to identify its purpose. To keep writing the
config as easy as possible the site configuration moved one level up and
now resides directly in config.yaml.

Additionally on writing top level keys that are not written by the GUI
will be kept, so extension authors can add additional configuration to
the site object without needing to configure the GUI (as that is hardly
extensible in the current state). Array keys // top level objects that
are configurable via GUI are however overwritten completely.

Resolves: #86323
Releases: master
Change-Id: I6100dbf53b0953e593a6156837de1fe013a74fae
Reviewed-on: https://review.typo3.org/58341
Reviewed-by: Tobi Kretschmann <tobi@tobishome.de>
Tested-by: Tobi Kretschmann <tobi@tobishome.de>
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>