[TASK] Remove a common set of false positives from extension scanner
The patch removes a set of common methods and properties from
the extension scanner to show less weak matches:
* method init()
* class properties ->config, ->tmpl, ->id, ->returnUrl
Remember this only kicks methods and properties that are so common
that many extensions use them and there are more false positives
than actual matches.
Remember a green file does not mean nothing breaks and a red file does not
mean nothing works. It is not goal of the extension scanner to find
everything, and matches do not mean they are valid.
The above matcher configurations have only been removed since they 
use a naming that is often also used in extension. The places where they are
used in core are usually deep in the API and it is rather unlikely a match found
by the scanner does really mean the deprecated / removed method or property
rather than just a false positive due to the common naming.
Other matchers should still be kept, even if they show false
positives in specific cases.
It should not be a goal of an extension developer to make the
extension scanner green.

Change-Id: Ifc27974fe9e854a93661cd1814f044f50a552119
Resolves: #86340
Releases: master
Reviewed-on: https://review.typo3.org/58360
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>