[TASK] Improve install tool loading message
The first 'silent configuration update' ajax call in the install
tool does two things: As all others, it checks for a valid session,
if valid, it executes the update.
If session is not active, the 'enable install tool' file view is
loaded next. Not having a dedicated 'check session' call as first
step saves one ajax round trip.
The patch improves the status message to avoid confusion why the
silent configuration update is executed for login view.

Resolves: #86347
Related: #86347
Releases: master
Change-Id: Idfd1fbb449dca4acd641f86e42092d56246ee3a2
Reviewed-on: https://review.typo3.org/58365
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Markus Klein <markus.klein@typo3.org>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>