[TASK] Update INSTALL.md and README.md
The main goal of this patch is to update the neglected
INSTALL.md and README.md to represent the current
state of the requirements and to improve the overall
content of the two "entry" documentation files.

The install.md has been massively shortened and links to
the official guides were added to finally be able to have
one source of truth.

Resolves: #86487
Releases: master
Change-Id: Id61abf86a4113d25c940c63f1423fa341aa7e702
Reviewed-on: https://review.typo3.org/58388
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Susanne Moog <susanne.moog@typo3.org>
Tested-by: Susanne Moog <susanne.moog@typo3.org>