[BUGFIX] Resolve correct processed folder when on different storage
During generation of processed files a wrong processed folder
is evaluated when the processed folder resides on a different storage.

We now ensure to ask the correct storage for the processed folder.

Also ensure that correct storage is used for checking file existence
when original file and processed files are on different storages.

Fixes: #86399
Releases: master, 8.7
Change-Id: I254b9a0bf86a156ae135510a4439e87737ec1a29
Reviewed-on: https://review.typo3.org/58410
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Nicole Cordes <typo3@cordes.co>
Tested-by: Nicole Cordes <typo3@cordes.co>