[TASK] Add informational upgrade wizard for argon2i
This adds a dummy wizard to remind admins during upgrade
to check the live system for argon2i support if the local
instance uses it, or to select a different hash algorithm.
Having this wizard gives this information to admins early
in the upgrade phase, so they have time to check the live
system or to select a different mechanism before too many
passwords have been upgraded.

Resolves: #86402
Releases: master
Change-Id: I2b1f75ecf079dc2e29d2675dda558c79b67f77e0
Reviewed-on: https://review.typo3.org/58411
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Susanne Moog <susanne.moog@typo3.org>
Tested-by: Susanne Moog <susanne.moog@typo3.org>
Reviewed-by: Frank Naegler <frank.naegler@typo3.org>
Tested-by: Frank Naegler <frank.naegler@typo3.org>