[BUGFIX] FailsafePackageManager: do not prevent package sorting
FailsafePackageManager prevents saving PackageStates.php in
failsafe mode by overwriting sortAndSavePackageStates(). That works
to prevent saving, but also prevents sorting when initially loading
packages when PackageStates.php is missing.

Currently this does not really produce problems, as the install
tool creates a new PackageStates.php later on and saves (and
implicitly sorts them) using forceSortAndPackageStates().

To fix the intermediate state we now only prevent saving,
but not the sorting of packages.

Releases: master
Resolves: #86408
Change-Id: I02ccb20653700e696d89b0a6ff39253f4f1ebc73
Reviewed-on: https://review.typo3.org/58416
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Frank Naegler <frank.naegler@typo3.org>
Tested-by: Frank Naegler <frank.naegler@typo3.org>
Reviewed-by: Susanne Moog <susanne.moog@typo3.org>
Tested-by: Susanne Moog <susanne.moog@typo3.org>