[TASK] Streamline phpdoc annotations in EXT:form
EXT:form uses @api and @internal phpDoc annotations. Going with a cleaner
approach with marking code as just @internal, and everything not annotated
as @internal is public, is the way to go, also the way PSR-5 is heading.

For EXT:form: Every PHP class that had nothing marked, is now @internal,
everything that was @api is now implicitly part of TYPO3 Core API.

Other areas will follow so the API for TYPO3 v9 LTS is clearer than ever.

Resolves: #86502
Releases: master
Change-Id: I2c1efdd11f0fbaa992315a90049eb109e2620c76
Reviewed-on: https://review.typo3.org/58521
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Ralf Zimmermann <ralf.zimmermann@tritum.de>
Tested-by: Ralf Zimmermann <ralf.zimmermann@tritum.de>