[TASK] Streamline phpdoc annotations in EXT:extbase
EXT:extbase uses @api and @internal phpDoc annotations. Going with a cleaner
approach with marking code as just @internal, and everything not annotated
as @internal is public, is the way to go, also the way PSR-5 is heading.

For EXT:extbase: Every PHP class that had nothing marked, is now @internal,
everything that was @api is now implicitly part of TYPO3 Core API.

On top, all license headers and @license annotations have been streamlined.

This means:
- TYPO3 Core's PHP classes area all public API by default
   unless marked as @internal or an extension class
- @api is not allowed anymore and will be restricted in
   the future from adding.
- @internal should be used for everything that should
   not be explicitly exposed as public API in the future.
- Everything under Tests/ is not part of TYPO3's Public API

Resolves: #86521
Releases: master
Change-Id: I83c5a27d9af001929142d2620600668ad0a84c92
Reviewed-on: https://review.typo3.org/58535
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Markus Klein <markus.klein@typo3.org>