[TASK] Add return type declaration to FileDumpController::dumpAction
For historical reasons FileDumpController::dumpAction was allowed to
return null. It didn't specify a nullable return type declaration because
it was not yed decided whether v9 will require php 7.2 or not.
With https://review.typo3.org/55585 FileDumpController was adapted to
return a PSR-7 Response but missed to update the function signature.
The nullable annotation is therefore removed now and the explicit
return type declaration is added.

Change-Id: I10e2cc60ab8e1f023c1beb6cf467bee1800971c8
Releases: master
Resolves: #86564
Reviewed-on: https://review.typo3.org/58574
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>