[BUGFIX] Send payload of image cropping wizard via POST
The image cropper wizard configuration can become very large, which
might break the wizard as the configuration is sent via query parameters
and "Request-URI Too Long" might kick in.

The payload is now sent via POST to bypass this issue. As our Modal API
is currently not capable of sending AJAX requests via POST, the logic
regarding the icon spinner is duplicated for the time being.

Resolves: #82225
Releases: master, 8.7
Change-Id: I7106b62fcc09101bc5147277225d1b8e89133d5c
Reviewed-on: https://review.typo3.org/58785
Tested-by: TYPO3com <no-reply@typo3.com>
Tested-by: Jonas Eberle <flightvision@googlemail.com>
Reviewed-by: Andreas Wolf <andreas.wolf@typo3.org>
Reviewed-by: Mathias Schreiber <mathias.schreiber@typo3.com>
Tested-by: Mathias Schreiber <mathias.schreiber@typo3.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>