[TASK] Always send 404 response on hidden pages
As a backend user, it is possible to preview a hidden page & record
in a workspace, however, without a user it is not possible.

Adding the tests also discovered that access to a "hidden" page should
result in a 404 instead of 403 response (even though the page
has access restrictions or not) in a live environment.

This change adds tests to ensure this functionality is always working.

Resolves: #92225
Resolves: #84098
Resolves: #86346
Releases: master, 10.4
Change-Id: I746717473bb93681c7b998d61c4f72eab4ec2ef3
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/58829
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>