[BUGFIX] Include DB queries with free mode records
With the new site handling, the overlayType for fetching
records in the frontend context is always set to
OVERLAYS_ON_WITH_FLOATING, which resolves to
the previous TypoScript settings:
- config.sys_language_overlay = 1
- styles.content.get.select.includeRecordsWithoutDefaultTranslation = 1

However, the place where "includeRecordsWithoutDefaultTranslation"
is evaluated does not check for the actual
Language Aspect of the Context API, which makes
it impossible to fetch records without "l10n_parent"
= (free mode), except for setting this option
explicitly on each TypoScript "select." query.

The patch moves the functionality into a separate method,
which allows to better unit test all options.

Resolves: #86627
Releases: master
Change-Id: I3a9005dcf13ddd26962da0d2ed930929c3f3b288
Reviewed-on: https://review.typo3.org/58859
Tested-by: TYPO3com <no-reply@typo3.com>
Tested-by: Chris Müller <typo3@krue.ml>
Reviewed-by: Chris Müller <typo3@krue.ml>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>