[BUGFIX] Enlarge scheduler object DB field
The serialized PHP object is stored within the database field
"serialized_task_object". When serializing this could lead to several
issues.

One particular issue is that the object has lots of data or dependencies,
it cannot be stored at all in this database field.

Although we know that this implementation detail of scheduler is a rather
bad idea, we can only fix this by enlarging the field from "blob" to
"mediumblob" for now.

Resolves: #87006
Releases: master, 8.7
Change-Id: I2b335f258fe18b494033bde28eff76f2e67d34bb
Reviewed-on: https://review.typo3.org/58953
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Riny van Tiggelen <info@online-gamer.nl>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Markus Klein <markus.klein@typo3.org>