[TASK] Do not show error when copying a record was successful
https://review.typo3.org/#/c/32356/ introduced the log message

The change was about skipping the direct copyRecord() call for
records that will be processed inside another copyRecord() call
for the record of the default language

Due to https://review.typo3.org/#/c/51070/ $overrideValues contains

['l10n_source' => 0]

When an element is copied inside an earlier copyRecord() call
the error log message is generated which leads to an error
flash message editors see when copying a page, but everything
(copying the records once) went fine.

A earlier call of copyRecord() can occur when using extension like
gridelements when copying a content element laying inside a
grid container element.

Resolves: #82032
Releases: master, 9.5, 8.7
Change-Id: I0d5be8e8920852a0e0c5a5ee93a67f9a6426e941
Reviewed-on: https://review.typo3.org/59234
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Mathias Brodala <mbrodala@pagemachine.de>
Tested-by: Mathias Brodala <mbrodala@pagemachine.de>