[TASK] Remove no_cache=1 from preview url generation
The usage of no_cache should be avoided. If still required it can be added
by using the setting additionalGetParameters.

As the functionality goes into the PageRouter (effectively adds cHash),
and it's possible that "no_cache" is completely disabled, it is not
necessary to be set. This was not the case in v9 as Site Handling
was not available everywhere.

On top, if an editor is actually previewing the content on the page
(via EditDocumentController "save + view") DataHandler is actually
flushing the cache of the page/record associated
(e.g. cacheTags for records), which is much more recommended.

Resolves: #87347
Releases: master
Change-Id: I2bcedb4f92231329234684c9daae2843d08ec29d
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/59362
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Benjamin Franzke <bfr@qbus.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benjamin Franzke <bfr@qbus.de>