[BUGFIX] Allow to access pages with "index.php?id=123"
Currently, the SiteMatcher checks for "?id=123&L=123" for special
treatment when Site Handling is enabled.

However, it should be possible to only access
www.example.com/index.php?id=123, without the "L" parameter,
for pages with Site-handling.

Without the patch, the language is not resolved, and a redirect
to /fr/ (or whatever default language prefix is given) is made.

Tests are adapted so IDs with ?id=1000 are not redirected
anymore.

Resolves: #87412
Releases: master, 9.5
Change-Id: Ied3df51359346996c6a889576b95516c008b734f
Reviewed-on: https://review.typo3.org/59414
Tested-by: TYPO3com <noreply@typo3.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>