[!!!][TASK] Drop TBE_TEMPLATE variable
The global variable TBE_TEMPLATE is not
necessary anymore, and is not available anymore.

The change is still ugly, as TBE_TEMPLATE
initializes the PageRenderer object
with necessary information, which should
become built inside PageRenderer for BE. This has been
moved to Backend's RequestHandler where this can be
abstracted further at a later point.

Resolves: #87567
Releases: master
Change-Id: If1cb184740f1e99c0aeca5ae19c85899d56f3f9a
Reviewed-on: https://review.typo3.org/59492
Tested-by: TYPO3com <noreply@typo3.com>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Daniel Gorges <daniel.gorges@b13.de>
Tested-by: Daniel Gorges <daniel.gorges@b13.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>