[TASK] Speed up ReST file reference check
There is a test related check for the extension scanner
to verify all referenced ReST files actually exist. It
is active as part of the 'integrity' job in bamboo.
The patch adds a variable to suppress checking the same
ReST file name multiple times if it is referenced in many
matchers. In practice, this speeds up the scanning process
drastically, on my local machine with current master from
more than a minute to less than ten seconds.

Resolves: #87493
Releases: master, 9.5
Change-Id: Iaa19add06ad9eeda9d36e3e204c1a671368b5f29
Reviewed-on: https://review.typo3.org/59495
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: TYPO3com <noreply@typo3.com>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>