[TASK] Use PSR-15 middlewares in install tool
Remove the old concept of priority-based HTTP request handlers,
and use statically configured PSR-15 middlewares instead.
Tranform the priority based request handlers into a set of
chained middlewares.

In case none of the middlewares handles the request a
NotFoundRequestHandler returns a 404 response
(replacing the previous exception that was thrown in that case).
This scheme of using a 404 handler as "last" request handler is
not new, it is also used in zend expressive (a PSR-15 framework).

The TYPO3 RequestHandlerInterface is marked @internal and was set
to be removed in v9, therefore we drop it now and also remove its
implementation from the frontend and backend request handlers.

Releases: master
Resolves: #87516
Change-Id: Ibbd6a0e6c4ff8aa55b4367c9e4c07e279595dbfb
Reviewed-on: https://review.typo3.org/59541
Tested-by: TYPO3com <noreply@typo3.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>