[SECURITY] Avoid creation of backend users without password
When using FormEngine it is possible to create a Backend User
without setting a password (or username), which could lead to
issues when using third-party authentication providers.

A hook within DataHandler ensures to set a random username
and/or password if the data is handed into DataHandler without
proper data. Besides that new backend users are disabled per
default and have to be enable manually.

Resolves: #80269
Releases: master, 9.5, 8.7
Security-Commit: 09b19dc181a565ca4a237f96747c0c808eb1c11b
Security-Bulletin: TYPO3-CORE-SA-2019-002
Change-Id: If4fb1e05c5dd8018077daa0c2a47779b2ca14342
Reviewed-on: https://review.typo3.org/59521
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>
Tested-by: Oliver Hader <oliver.hader@typo3.org>