[SECURITY] Avoid creation of backend users without password
When using FormEngine it is possible to create a Backend User
without setting a password (or username), which could lead to
issues when using third-party authentication providers.

A hook within DataHandler ensures to set a random username
and/or password if the data is handed into DataHandler without
proper data. Besides that new backend users are disabled per
default and have to be enable manually.

Resolves: #80269
Releases: master, 9.5, 8.7
Security-Commit: f8a9edfed26ad48d13564ea99f27e0846671841c
Security-Bulletin: TYPO3-CORE-SA-2019-002
Change-Id: Ic1d84010717e3ac056f447fd373b31bbce8f65c6
Reviewed-on: https://review.typo3.org/59535
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>
Tested-by: Oliver Hader <oliver.hader@typo3.org>