[TASK] bamboo: db dependency loop needs break condition
Functional and acceptance tests need a db up and running.
The according container is started as dependency and a
loop delays further execution until the database connected
to its network port. This sometimes goes wrong, for
instance mssql in rare conditions does not come up. This
leads to the loop running "forever", consuming the
executing bamboo agent and the job never finishs.
The patch adds an additional break condition: If db did
not connect to the port after a minute, the job now fails.

Resolves: #87549
Releases: master, 9.5, 8.7
Change-Id: Idc45f57d9cfd2d390cea6bd5319adaf4158340df
Reviewed-on: https://review.typo3.org/59569
Tested-by: TYPO3com <noreply@typo3.com>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>