[BUGFIX] Resolve site correctly in nested site structure
Due to a bug in checking for $result[t3ver_oid] ?? $result[uid]
the uid is never used, but only the t3ver_oid. This value could be empty
and then "0" is used. This means, the page record to be edited
in FormEngine is treated as "NullSite" = Page without Site Configuration.

Resolves: #87572
Releases: master, 9.5
Change-Id: I1c4175f61a9c5df3dd3a392c2f636112cfbdb350
Reviewed-on: https://review.typo3.org/59579
Tested-by: TYPO3com <noreply@typo3.com>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Daniel Gorges <daniel.gorges@b13.de>
Tested-by: Daniel Gorges <daniel.gorges@b13.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>