[BUGFIX] Use real finisher identifier in form element variants
Use the real finisher identifiers as they are configured in the
form definition instead the finisher implementation class name.
This requires a change of the public AbstractFinisher implementation
which implements the FinisherInterface. The interface is not changed,
as this would be a breaking change. Therefore, the existence of the
new method is checked. If it does not exist in a finisher
implementation, a fallback to the previous behavior happens.

Releases: master, 9.5
Resolves: #87615
Change-Id: I962502f0601845a5f9793751e87c97b3b902caa1
Reviewed-on: https://review.typo3.org/59606
Tested-by: TYPO3com <noreply@typo3.com>
Reviewed-by: Susanne Moog <susanne.moog@typo3.org>
Tested-by: Susanne Moog <susanne.moog@typo3.org>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>