[BUGFIX] Do not treat all classes as entities in ClassSchema
The ClassSchema class has a property $modelType which indicates
that a class is either an entity or a value object. This property
has a default value that reflected the state of an entity.

This means, that all classes, class schemata are created for are
considered entities which then leads to the execution of further
entity related analysis code, such as looking for specific doc
blocks and/or annotations.

While this isn't an issue functionality wise, it can be in the future.
Also, the superfluous cpu cycles can and should be spared.

Releases: master, 9.5
Resolves: #87632
Change-Id: If4bc79b9761ee28ffd9e2f25924588140310c0fa
Reviewed-on: https://review.typo3.org/c/59792
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>